Insert doc comments when typing via LSP (#46301)
